partition,bootloader: rename 'partition' package to 'bootloader'
The 'partition' package is actually about bootloaders, so 'bootloader' name
seems more fitting.
