+++ boot/kernel_os.go
@@ -24,8 +24,8 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/snapcore/snapd/bootloader"
 	"github.com/snapcore/snapd/logger"
-	"github.com/snapcore/snapd/partition"
 	"github.com/snapcore/snapd/release"
 	"github.com/snapcore/snapd/snap"
 )
@@ -33,14 +33,14 @@ import (
 // RemoveKernelAssets removes the unpacked kernel/initrd for the given
 // kernel snap.
 func RemoveKernelAssets(s snap.PlaceInfo) error {
-	bootloader, err := partition.FindBootloader()
+	loader, err := bootloader.FindBootloader()
 	if err != nil {
 		return fmt.Errorf("no not remove kernel assets: %s", err)
 	}
 
 	// remove the kernel blob
 	blobName := filepath.Base(s.MountFile())
-	dstDir := filepath.Join(bootloader.Dir(), blobName)
+	dstDir := filepath.Join(loader.Dir(), blobName)
 	if err := os.RemoveAll(dstDir); err != nil {
 		return err
 	}
@@ -56,18 +56,18 @@ func ExtractKernelAssets(s *snap.Info, snapf snap.Container) error {
 		return fmt.Errorf("cannot extract kernel assets from snap type %q", s.Type)
 	}
 
-	bootloader, err := partition.FindBootloader()
+	loader, err := bootloader.FindBootloader()
 	if err != nil {
 		return fmt.Errorf("cannot extract kernel assets: %s", err)
 	}
 
-	if bootloader.Name() == "grub" {
+	if loader.Name() == "grub" {
 		return nil
 	}
 
 	// now do the kernel specific bits
 	blobName := filepath.Base(s.MountFile())
-	dstDir := filepath.Join(bootloader.Dir(), blobName)
+	dstDir := filepath.Join(loader.Dir(), blobName)
 	if err := os.MkdirAll(dstDir, 0755); err != nil {
 		return err
 	}
@@ -104,7 +104,7 @@ func SetNextBoot(s *snap.Info) error {
 		return fmt.Errorf("cannot set next boot to snap %q with type %q", s.SnapName(), s.Type)
 	}
 
-	bootloader, err := partition.FindBootloader()
+	bootloader, err := bootloader.FindBootloader()
 	if err != nil {
 		return fmt.Errorf("cannot set next boot: %s", err)
 	}
@@ -154,7 +154,7 @@ func ChangeRequiresReboot(s *snap.Info) bool {
 		return false
 	}
 
-	bootloader, err := partition.FindBootloader()
+	bootloader, err := bootloader.FindBootloader()
 	if err != nil {
 		logger.Noticef("cannot get boot settings: %s", err)
 		return false
@@ -187,7 +187,7 @@ func ChangeRequiresReboot(s *snap.Info) bool {
 // InUse checks if the given name/revision is used in the
 // boot environment
 func InUse(name string, rev snap.Revision) bool {
-	bootloader, err := partition.FindBootloader()
+	bootloader, err := bootloader.FindBootloader()
 	if err != nil {
 		logger.Noticef("cannot get boot settings: %s", err)
 		return false
