+++ bootloader/androidbootenv/androidbootenv_test.go
@@ -25,7 +25,7 @@ import (
 
 	. "gopkg.in/check.v1"
 
-	"github.com/snapcore/snapd/partition/androidbootenv"
+	"github.com/snapcore/snapd/bootloader/androidbootenv"
 )
 
 // Hook up check.v1 into the "go test" runner
