+++ image/image.go
@@ -32,9 +32,9 @@ import (
 	"github.com/snapcore/snapd/asserts/snapasserts"
 	"github.com/snapcore/snapd/asserts/sysdb"
 	"github.com/snapcore/snapd/boot"
+	"github.com/snapcore/snapd/bootloader"
 	"github.com/snapcore/snapd/dirs"
 	"github.com/snapcore/snapd/osutil"
-	"github.com/snapcore/snapd/partition"
 	"github.com/snapcore/snapd/release"
 	"github.com/snapcore/snapd/snap"
 	"github.com/snapcore/snapd/snap/squashfs"
@@ -665,7 +665,7 @@ func setupSeed(tsto *ToolingStore, model *asserts.Model, opts *Options, local *l
 
 	if !opts.Classic {
 		// now do the bootloader stuff
-		if err := partition.InstallBootConfig(opts.GadgetUnpackDir); err != nil {
+		if err := bootloader.InstallBootConfig(opts.GadgetUnpackDir); err != nil {
 			return err
 		}
 
@@ -690,7 +690,7 @@ func setBootvars(downloadedSnapsInfoForBootConfig map[string]*snap.Info, model *
 	// Set bootvars for kernel/core snaps so the system boots and
 	// does the first-time initialization. There is also no
 	// mounted kernel/core/base snap, but just the blobs.
-	bootloader, err := partition.FindBootloader()
+	loader, err := bootloader.FindBootloader()
 	if err != nil {
 		return fmt.Errorf("cannot set kernel/core boot variables: %s", err)
 	}
@@ -737,7 +737,7 @@ func setBootvars(downloadedSnapsInfoForBootConfig map[string]*snap.Info, model *
 			m[bootvar] = name
 		}
 	}
-	if err := bootloader.SetBootVars(m); err != nil {
+	if err := loader.SetBootVars(m); err != nil {
 		return err
 	}
