+++ overlord/snapstate/booted_test.go
@@ -29,11 +29,11 @@ import (
 	. "gopkg.in/check.v1"
 
 	"github.com/snapcore/snapd/boot/boottest"
+	"github.com/snapcore/snapd/bootloader"
 	"github.com/snapcore/snapd/dirs"
 	"github.com/snapcore/snapd/overlord"
 	"github.com/snapcore/snapd/overlord/snapstate"
 	"github.com/snapcore/snapd/overlord/state"
-	"github.com/snapcore/snapd/partition"
 	"github.com/snapcore/snapd/release"
 	"github.com/snapcore/snapd/snap"
 	"github.com/snapcore/snapd/snap/snaptest"
@@ -68,7 +68,7 @@ func (bs *bootedSuite) SetUpTest(c *C) {
 	bs.bootloader = boottest.NewMockBootloader("mock", c.MkDir())
 	bs.bootloader.BootVars["snap_core"] = "core_2.snap"
 	bs.bootloader.BootVars["snap_kernel"] = "canonical-pc-linux_2.snap"
-	partition.ForceBootloader(bs.bootloader)
+	bootloader.ForceBootloader(bs.bootloader)
 
 	bs.fakeBackend = &fakeSnappyBackend{}
 	bs.o = overlord.Mock()
@@ -94,7 +94,7 @@ func (bs *bootedSuite) TearDownTest(c *C) {
 	snapstate.Model = nil
 	release.MockOnClassic(true)
 	dirs.SetRootDir("")
-	partition.ForceBootloader(nil)
+	bootloader.ForceBootloader(nil)
 }
 
 var osSI1 = &snap.SideInfo{RealName: "core", Revision: snap.R(1)}
